0.0.13

+++ src/operations/channelInvitation/deleteChannelInvitation.ts
@@ -1,11 +1,16 @@
 import db from '../../db/db.js';
 import { ModelType, MutationType } from '../../enums.js';
+import libData from '../../helpers/libData.js';
 import { ChannelInvitation } from '../../types/models/ChannelInvitation.js';
 import { MutationResult } from '../../types/MutationResult.js';
 
 const deleteChannelInvitation = async (
   id: string,
 ): Promise<MutationResult<ChannelInvitation>> => {
+  if (!libData.isInitialized()) {
+    throw new Error('not-initialized');
+  }
+
   try {
     await db.delete(id, ModelType.ChannelInvitation);
